// frontend/src/types/clips.ts
interface HasId {
  _id: string;
}

export interface MetaMeta {
  created?: string;
  createdBy?: string;
  updated?: string;
  updatedBy?: string;
}
// ...
export interface Annotation extends HasId, MetaMeta {
  type: 'Annotation';
  clip: string;
  element?: string;
  relation: Relation;
  target: string;
  quotes: string;
  date?: string;
  role?: string;
  instrument?: string;
  metaDate?: string;
  timecodeStart?: number;
  timecodeEnd?: number;
  confidence?: 'certain' | 'unsure' | 'maybe' | ''; // tbd
  // review Franziska uncertainty stuff before making decision here
  attribution?: string;
  // attribution?: {
  //   zoteroId: string;
  //   locator?: string;
  //   attributionComment?: string;
  // };
  comment?: string; // we want comment threads, right?
  fromPlatformMetadata?: boolean;
  interpretative?: boolean;
  constitutedBy?: string[]; // annotation ids... autocomplete w relation and target/data
  refersTo?: string; // annotation id ("Der Handshake ist ein symbolischer Akt")
  // flags: Flag[];
  notablyAbsent?: boolean;
}
// ...
interface ClipConstruct extends HasId, MetaMeta {
  clip: string;
  label: string;
  description: string;
  timecodes: Array<[number, number]>;
}

export interface MusicBasic extends ClipConstruct {
  type: 'Music';
}

export interface Music extends MusicBasic {
  annotations: Record<MusicRelation, string[]>;
}

export interface SpeechBasic extends ClipConstruct {
  type: 'Speech';
}

export interface Speech extends SpeechBasic {
  annotations: Record<SpeechRelation, string[]>;
}

export interface NoiseBasic extends ClipConstruct {
  type: 'Noise';
}

export interface Noise extends NoiseBasic {
  annotations: Record<NoiseRelation, string[]>;
}

export interface PictureBasic extends ClipConstruct {
  type: 'Picture';
}

export interface Picture extends PictureBasic {
  annotations: Record<PictureRelation, string[]>;
}

// interface SegmentContains {
//   [elementId: string]: { [relation: string]: SegmentMember[] };
// }

export interface SegmentMember {
  added: string;
  addedBy: string;
  annotation: string; // annotation id
}

export type ElementType = 'Segment' | 'Music' | 'Speech' | 'Noise' | 'Picture';

export interface ClipElement extends ClipConstruct {
  type: ElementType;
  clip: string;
}

export interface Segment extends ClipElement {
  type: 'Segment';
  timecodes: Array<[number, number]>;
  segmentContains: SegmentMember[];
}